Changes 64 bits family to fast 32 bits

+++ entityx/Event.h
@@ -26,7 +26,7 @@ namespace entityx {
 /// Used internally by the EventManager.
 class BaseEvent {
  public:
-  typedef uint64_t Family;
+  typedef size_t Family;
 
   virtual ~BaseEvent() {}
